feat(menu): integrate advanced graph algorithms into legacy, TUI, and benchmark menus
