+++ src/main.c
@@ -1,3 +1,4 @@
+#include "advanced_graph_algorithms/advanced_graph_algorithms.h"
 #include "advanced_sorting_algorithms/advanced_sorting.h"
 #include "backtracking/backtracking.h"
 #include "benchmark.h"
@@ -52,8 +53,10 @@ void run_legacy_menu()
             "click 14 for process synchronization algorithms demo\n"
             "click 15 for setting animation speed (by default 2s)\n"
             "click 16 for algorithm benchmarking and profiling demo\n"
+            "click 17 for advanced graph algorithms (scc / max flow / bipartite matching / "
+            "eulerian path) demo\n"
             "enter choice : ",
-            1, 16 // limits
+            1, 17 // limits
         );
 
         if (status == -111)
@@ -117,6 +120,9 @@ void run_legacy_menu()
             case 16:
                 display_header("Algorithm Benchmarking & Profiling");
                 benchmark_menu_demo();
+                break;
+            case 17:
+                advanced_graph_algorithms_demo();
                 break;
         }
     }
